Fix a bug due to parsing order

+++ src/Ext/Rules/Body.h
@@ -29,7 +29,7 @@ public:
 		Valueable<int> RadApplicationDelay_Building;
 		PhobosFixedString<32u> MissingCameo;
 
-		Valueable<float> JumpjetCrash;
+		Valueable<double> JumpjetCrash;
 		Valueable<bool> JumpjetNoWobbles;
 
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
@@ -37,7 +37,7 @@ public:
 			, Pips_Shield_Buildings({ -1,-1,-1 })
 			, RadApplicationDelay_Building(0)
 			, MissingCameo("xxicon.shp")
-			, JumpjetCrash(5.0f)
+			, JumpjetCrash(5.0)
 			, JumpjetNoWobbles(false)
 		{ }
